# Obsevatoire_de_recherche/src/Backend/laravel10/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for my application.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/ressources/filieres": {
            "get": {
                "tags": [
                    "Filieres"
                ],
                "summary": "Get list of all filieres",
                "operationId": "4f73faa12b9c70cc0a73e2bd525531b5",
                "responses": {
                    "200": {
                        "description": "A list of filieres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblFiliere"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Filieres"
                ],
                "summary": "Create a new filiere",
                "operationId": "c4a71b949c6881743e3c63b0f118fbc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblFiliere"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Filiere created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblFiliere"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ressources/filieres/{id}": {
            "get": {
                "tags": [
                    "Filieres"
                ],
                "summary": "Get a filiere by ID",
                "operationId": "bf75d13638ad78c09fa151c169f7c5e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filiere details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblFiliere"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Filiere not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Filieres"
                ],
                "summary": "Update a filiere",
                "operationId": "aa86282a3c76fceabcad172af33a947c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblFiliere"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filiere updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblFiliere"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Filiere not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Filieres"
                ],
                "summary": "Delete a filiere",
                "operationId": "0654372dd424ddb5c879ce89503d5cdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Filiere deleted"
                    },
                    "404": {
                        "description": "Filiere not found"
                    }
                }
            }
        },
        "/api/ressources/niveaux": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Get list of all niveaux",
                "operationId": "b76da7f932d8781753b2f59b7c200bfe",
                "responses": {
                    "200": {
                        "description": "A list of niveaux",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblNiveau"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Create a new niveau",
                "operationId": "a5e744f0e0c69409ca61a3c701cad19b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblNiveau"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Niveau created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblNiveau"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/ressources/niveaux/{id}": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Get a niveau by ID",
                "operationId": "d80a2b2135d8766c506b7cfccd9eefd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Niveau details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblNiveau"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Niveau not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Update a niveau",
                "operationId": "0895c8a582ccc152ce5cab12f6b04988",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TblNiveau"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Niveau updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TblNiveau"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Niveau not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Delete a niveau",
                "operationId": "0fd7754e9557b367ff1be550955b013a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Niveau deleted"
                    },
                    "404": {
                        "description": "Niveau not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TblFiliere": {
                "title": "TblFiliere",
                "required": [
                    "nom_fil",
                    "tbl_faculte_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the filiere",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_fil": {
                        "description": "Name of the filiere",
                        "type": "string"
                    },
                    "tbl_faculte_id": {
                        "description": "ID of the related faculte",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "TblNiveau": {
                "title": "TblNiveau",
                "required": [
                    "code_niv"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the niveau",
                        "type": "integer",
                        "format": "int64"
                    },
                    "code_niv": {
                        "description": "Code of the niveau",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}